Removed unused NavBar import. Converted to a jsx file. Needed to resolve an issue when importing LoginPage. Moved App.test.js to the root of the project.
